refactor: move to separate files, fix restart game
